+++ modules/em8300_ioctl.c
@@ -32,7 +32,7 @@ int em8300_control_ioctl(struct em8300_s *em, int cmd, unsigned long arg)
 	em8300_overlay_screen_t ov_scr;
 	em8300_overlay_calibrate_t ov_cal;
 	em8300_attribute_t attr;
-	unsigned int safe_jiff = jiffies;
+	unsigned long safe_jiff = jiffies;
 
 	if (_IOC_DIR(cmd) != 0) {
 		len = _IOC_SIZE(cmd);
