+++ modules/em8300_spu.c
@@ -75,7 +75,7 @@ void em8300_spu_check_ptsfifo(struct em8300_s *em)
 ssize_t em8300_spu_write(struct em8300_s *em, const char * buf, size_t count, loff_t *ppos)
 {
 	int flags = 0;
-	unsigned int safe_jiff = jiffies;
+	unsigned long safe_jiff = jiffies;
 
 	if (!(em->sp_mode)) return 0;
 //	em->sp_ptsvalid=0;
